Respect the --verbose flag for `apm install`
`apm install [--verbose]` can now recognize the `--verbose` flag,
and send it to `npm` to form `npm install --verbose`.

This prints lots of useful information that can be used
to debug what happens when users try to install a package,
or when users try to install the dependencies for a package
during development.