meta: Disable prepare script (coffee compilation)
We are going to use the JS output directly now.

So, we want to make sure the JS code isn't accidentally overwritten by
the prepare script when doing `npm install` or `yarn install`.

The `build` script can still be run as `npm run build`, if compiling
the CoffeeScript files from the `src/` dir is needed... Such as when
merging new changes that may have landed in `master` branch while this
decaffeination work is ongoing. Code may have to be synced a few times
across the branches, with new changes decaffeinated and merged into
the JS output in `lib/`, which probably has to be done at least
somewhat manually -- I don't think that could be 100% automated.