script: New script to download Node source files
This will be used for fetching NodeJS source and headers, and other
related files, which are used for running certain tests
having to do with building native C/C++ modules.

(Updating the version of Node's source/headers we build against in the
tests has proven necessary from time to time, for reasons that aren't
entirely clear. It seems most likely to me that `node-gyp` has trouble
running on a recent NodeJS version but building against a much older
Node version, perhaps due to subtle changes in Node over time, and due
to assumptions made in node-gyp that the running version of Node will
have similar properties to the target version of Node, with those
assumptions failing to anticipate the actual scope of changes made to
newer Node compared to older versions?)

So, we do have to update these source/header files every do often.

Newer NodeJS versions are getting bigger and bigger, to the point
where checking in all the necessary files would amount to over 100
megabytes, and potentially triple the bandwitdth/disk space needed to
clone this repository. And it would be at least as bad all over again
if we even needed to bump these files to even newer Node later on.

To this day, one of the largest contributors to this repo's size was
the bump to Node v10.20.1 I did over a year ago. Small text files
compress incredibly efficiently, whereas binary files don't compress
anywhere near as well (like Node's source tarballs, for example).

GitHub also warns when you upload files over 50 megabytes,
whereas some of these tarballs are closer to 70 or 80 megabytes now.

So, this script ensures we can fetch & populate the files into this
repo's spec/fixtures dir just before running the tests.

If you've already run it once, this script  will simply check the
cryptographic hashes of the files, to make sure they were downloaded
correctly. If the hash is incorrect, it will attempt to redownload
at most once per already-existing file before giving up.

This should do a reasonable job at ensuring we reliably have these
files, without having to check them in to the repo in git, and without
wasting a lot of bandwidth and disk space for a simple clone of this
repo.

(Git submodules were another option, but their tooling is somewhat
clunky in my opinion, and it would just be punting the
"really large files" problem to somewhere else on GitHub. I think it's
much better to simply fetch these large files off the web as-needed,
and make some attempt to cache them/keep them around afterward, rather
than checking them in to any sort of main git repo or submodule repo.
So that's what this script does instead -- fetch files from the web.)