script: run ensure-fixtures script before tests
Runs the new script to make sure we have the needed node source files
(used by certain tests having to do with building native C/C++
modules), before tests start.

(These source files will be needed for the tests to pass once the old,
checked-in source files are deleted, and we switch to newer ones
that will be fetched on-demand from nodejs.org on the web.)

(The new files are too big to check in without bloating the repository
and causing a warning from GitHub -- some of the files are over 50MB.)

This commit also mildly re-factors the `test` script to break out
the `specs` into their own sub-script. Makes it easier to read the
shorter `test` script on one line. Also makes it so you can quickly
re-run the specs if you already have the repo prepared and the
.coffee files haven't been updated.