meta: .gitignore the new node-source fixtures
These new fixture source files are *not* meant to be checked in.
They should be fetched on-demand and cached/re-downloaded as needed.

(These source files are becoming way too big to check in to git each
time we need to update them to a newer Node version. They get bigger
with each Node version, as NodeJS continues to add new features.)