script: Rearrange filesToFetch to match dir order
This is the order the files are listed in Windows' `dir` command,
and the order they will be displayed when locally re-calculating
all the file hashes, assuming these ever need to be updated.

This also much closer matches the order the files show in Linux's `ls`
command, as well as in the `shasum` command available on Linux/macOS.

So in general, this should make copy-pasting the updated hashes
much less confusing for whoever has to do that next.