spec,script: Test against Electron v12.2.3
Test against Electron v12.2.3, not Node v18.12.1.

---

This commit introduces a `distUrl` variable to
"script/ensure-node-source-fixture-files.js", where we can specify
the Node or Electron dist URL we are grabbing our source files
and headers from, and so on...

and sets it to the official Electron headers/tarballs dist URL,
so we can test against Electron -- particularly the Electron version
that Pulsar uses.

Also updates `nodeVersion` for this script and the specs to "v12.2.3".

---

In effect: Test building native dummy modules against Electron 12,
not Node 18, during the relevant specs.