spec: Fix dist structure for node.lib files
This is the new folder structure for the node.lib files
(of various arches), which are used only on Windows.

(This directory structure change apparently happened all the way back
with the iojs fork (version 1.x in 2015) and carried through to NodeJS
when iojs was eventually used as the basis of NodeJS 4.x and newer.)

See the location of the `node.lib` files at:
- https://nodejs.org/dist/v0.11.16/node.lib (old x86 path)
- https://nodejs.org/dist/v0.11.16/x64/node.lib (old x64 path)
...
- https://iojs.org/dist/v1.0.0/win-x64/iojs.lib (new x64 path)
- https://nodejs.org/dist/v4.0.0/win-x64/node.lib (new x64 path)
- https://nodejs.org/dist/v19.3.0/win-x64/node.lib (new x64 path)
- https://nodejs.org/dist/v19.3.0/win-x86/node.lib (new x86 path)

Since this directory structure change happened, newer node-gyp
requests the files from the updated paths used by this commit,
not the old paths (the ones from all the way before version 1.0
of NodeJS or Electron were ever released! ~8 years ago!)

So we need to serve node-gyp's requests pointing to these new paths
for the spec fixture files to actually serve their purpose,
and for compilation to work on Windows in these tests (if the user
hasn't already downloaded/cached them from a previous node-gyp run.)

The tests still don't work on Windows right now, for unknown reasons,
but it does appear from my testing to have _something_ to do with
native module compilation and node-gyp?

It's good to be actually providing these files as are needed for
those specs under Windows, regardless of other reasons for Windows-
specific test failures. One less problem makes it likelier that
we get everything resolved, however unlikely that may still be.