spec: Move 'nodeVersion' to a JSON file, D.R.Y
This version string is part of the filenames and paths (dist server
directory structure) used by node-gyp when fetching headers and
dynamic libraries, which are used to build native C/C++ addons against
NodeJS or Electron for use in various JS modules.

So, we have to have the version number of these files in order to run
a working "dist server" serving these files for use in the specs.

We can put this version string in a JSON file that is read from
all the spec files that need it, rather than hard-coding "v12.2.3"
several times per file, across several different spec files.

This version string should always match the version of the headers
tarball, node.lib, and SHASUMS256.txt files under `spec/fixtures/...`.
Update the version string and the files together, at the same time.