bin: Add 'ppm' and 'ppm.cmd' bins
Makes it easier to integrate ppm in various contexts, such as on the
user's PATH to be run from the CLI, without having to run it as 'apm'.

Leaving the old 'apm' and 'apm.cmd' bins in case these are useful for
troubleshooting during the transition to 'ppm'-named bins, or in case
some edge case requires them.

They're two small text files, I think we can afford to keep ones with
either name, at least for now, if not forever. We'll see what's best.