script: Adjust 'ppm' bin file perms on Windows
We just added a 'ppm' bin, better treat it the same as the 'apm' bin.

Apparently the file permissions for these binaries get easily
lost in translation when synced between Linux and Windows and back...

So, we explicitly set correct file permissions for the bins during
the postinstall script on Windows. Otherwise, the published bins
are somewhat broken (not easily runnable) for Linux users
when published from a Windows machine to the npm package registry
and then downloaded from there onto a Linux machine.

(Pulsar team doesn't publish ppm to the npm package registry,
but you never know when we might want to start doing that,
like they used to do with apm?)

(We might as well keep this "legacy code" functional,
ensuring parity for all relevant files it pertains to,
unless we're going to outright delete said "legacy code" entirely...)