src: Stop pinging backend during package uninstalls
So, we stopped *using* these pings on our implementation of the
backend, since it doesn't really make sense to decrement
the install count when someone uninstalls... And it may or may not
have simpoly been telemetry when the O.G. Atom backend was expecting
this to be an authed action.

Since we don't need to auth or ping the backend for uninstalls...
Just delete the code that does it!