deps: Bump node-gyp to latest ^10.2.0 for Python 3.12 compat
Includes an edition of nodejs/gyp-next which vendors Python 'packaging'
library and moves away from 'distutils', which used to be shipped with
Python, but was deprecated, and then dropped starting in Python 3.12.

---

This solves an error -- Python 3.12 users without 'distutils' (or
'setuptools' which provides 'distutils') on their systems would get
an error like the following when trying to build native C/C++ code:

```
    from distutils.version import StrictVersion
ModuleNotFoundError: No module named 'distutils'
```

Newer gyp-next 0.16.0+ (and by extension node-gyp 10+) avoids the error
by not depending on the library that was dropped from being included
"out of the box" with Python as of Python 3.12. (It vendors 'packaging'
module to do its version string parsing, rather than relying on the
now-missing 'distutils'.)