Infer an `ATOM_BASE_NAME` environment variable…
…from the name of the executable.

Used to keep separate release channels (`pulsar` vs. `pulsar-next`) with separate home folders.

In order to allow people to run PulsarNext (Pulsar on Electron 30+) alongside the stable version of Pulsar, we must create a separate app with a separate name and a separate `ATOM_HOME`!

This requires that the version of `ppm` bundled with PulsarNext also understand which release channel it's using so that it can install packages into the proper location.

For the “next” release channel, our current way of pulling this off is to write the `ppm` binary to `ppm-next` instead of `ppm`. The command won't conflict with an existing `ppm` in the user's `PATH`, plus PPM itself will be able to infer the release channel from its own name.